# openapi.yaml
openapi: 3.0.3
info:
  title: LeanAutoLinks REST API
  version: 1.0.0
  description: |
    The LeanAutoLinks REST API provides programmatic access to all plugin functionality:
    managing linking rules, monitoring the background processing queue, viewing applied
    links and statistics, managing exclusions, and monitoring performance and cache health.

    ## Authentication

    All endpoints require WordPress Application Passwords using HTTP Basic Auth.
    The authenticated user must have the `manage_options` capability.

    ```bash
    curl -u "admin:xxxx xxxx xxxx xxxx xxxx xxxx" \
      https://example.com/wp-json/leanautolinks/v1/rules
    ```

    Generate an Application Password at **Users > Profile > Application Passwords** in
    your WordPress admin dashboard.

    ## Pagination

    List endpoints follow WordPress REST API conventions. Use `page` and `per_page`
    query parameters to paginate results. Responses include the following headers:

    | Header | Description |
    |--------|-------------|
    | `X-WP-Total` | Total number of items matching the query |
    | `X-WP-TotalPages` | Total number of pages |

    ## Error Handling

    All error responses use a consistent JSON structure:

    ```json
    {
      "code": "invalid_parameter",
      "message": "The rule_type parameter must be one of: internal, affiliate, entity.",
      "data": { "status": 400, "param": "rule_type" }
    }
    ```

    ## Rate Limiting

...

    ## Versioning

...
  contact:
...
servers:
...
security:
  - basicAuth: []
...
paths:
...
  /applied:
    get:
      operationId: listAppliedLinks
      summary: List applied links
      description: |
        Returns a paginated list of links that the engine has inserted into
        post content. You can filter by post ID (to see all links in one post)
        or by rule ID (to see everywhere a rule has been applied).
      tags: [Applied Links]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: post_id
          in: query
          description: Filter by WordPress post ID.
          required: false
          schema:
            type: integer
            format: int64
        - name: rule_id
          in: query
          description: Filter by rule ID.
          required: false
          schema:
            type: integer
            format: int64
        - name: orderby
          in: query
          description: Field to order results by.
          required: false
          schema:
            type: string
            enum: [id, post_id, rule_id, applied_at]
            default: applied_at
        - name: order
          in: query
          description: Sort direction.
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: A paginated list of applied links.
          headers:
            X-WP-Total:
              $ref: '#/components/headers/X-WP-Total'
            X-WP-TotalPages:
              $ref: '#/components/headers/X-WP-TotalPages'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppliedLink'
              example:
                - id: 10234
                  post_id: 8234
                  rule_id: 1
                  keyword: inteligencia artificial
                  target_url: https://example.com/glosario/inteligencia-artificial/
                  applied_at: "2026-03-24T10:15:02"
                - id: 10235
                  post_id: 8234
                  rule_id: 5
                  keyword: Y Combinator
                  target_url: https://example.com/actores/y-combinator/
                  applied_at: "2026-03-24T10:15:02"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /applied/stats:
    get:
      operationId: getAppliedStats
      summary: Get applied link statistics
      description: |
        Returns aggregated statistics about applied links: total count,
        breakdown by rule type, and the posts with the most applied links.
      tags: [Applied Links]
      responses:
        '200':
          description: Applied link statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppliedStats'
              example:
                total_links: 48210
                by_rule_type:
                  internal: 28430
                  entity: 15200
                  affiliate: 4580
                posts_with_links: 14320
                avg_links_per_post: 3.37
                top_linked_posts:
                  - post_id: 1204
                    post_title: "Guia completa de inteligencia artificial para startups"
                    link_count: 12
                  - post_id: 3892
                    post_title: "El ecosistema de venture capital en Latinoamerica"
                    link_count: 11
                  - post_id: 5017
                    post_title: "Herramientas de productividad para emprendedores"
                    link_count: 10
                top_rules:
                  - rule_id: 1
                    keyword: inteligencia artificial
                    applied_count: 4230
                  - rule_id: 15
                    keyword: startup
                    applied_count: 3890
                  - rule_id: 42
                    keyword: venture capital
                    applied_count: 2105
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  # ---------------------------------------------------------------------------
  # Exclusions
  # ---------------------------------------------------------------------------
...
  /performance/summary:
    get:
      operationId: getPerformanceSummary
      summary: Get aggregated performance metrics
      description: |
        Returns aggregated performance metrics including average processing
        time per post, cache hit rate, queue depth, and throughput statistics.
        Metrics are computed over the last 24 hours by default, configurable
        via the `period` parameter.
      tags: [Performance]
      parameters:
        - name: period
          in: query
          description: |
            Time period for metric aggregation. Accepted values: `1h`, `6h`,
            `24h`, `7d`, `30d`.
          required: false
          schema:
            type: string
            enum: ["1h", "6h", "24h", "7d", "30d"]
            default: "24h"
      responses:
        '200':
          description: Aggregated performance metrics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceSummary'
              example:
                period: "24h"
                processing:
                  total_posts_processed: 1420
                  avg_duration_ms: 187.4
                  p95_duration_ms: 432.0
                  max_duration_ms: 890.0
                  avg_rules_checked: 1024
                  avg_links_applied: 3.2
                  avg_memory_kb: 8420
                  max_memory_kb: 24100
                cache:
                  hit_rate: 0.94
                  hits: 45230
                  misses: 2870
                  estimated_memory_mb: 148.5
                queue:
                  pending: 12
                  processing: 3
                  done: 14832
                  failed: 2
                  estimated_drain_seconds: 45
                throughput:
                  posts_per_minute: 14.8
                  sustained_posts_per_hour: 888
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /performance/log:
    get:
      operationId: getPerformanceLog
      summary: Get detailed performance log
      description: |
        Returns individual performance log entries. Each entry represents a
        single processing event (one post processed) with detailed timing,
        memory, and result metrics.
      tags: [Performance]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: event_type
          in: query
          description: Filter by event type.
          required: false
          schema:
            type: string
            enum:
              - process_post
              - process_failed
              - cache_hit
              - cache_miss
              - cache_flush
              - cache_warm
              - bulk_start
              - bulk_complete
        - name: post_id
          in: query
          description: Filter log entries for a specific post.
          required: false
          schema:
            type: integer
            format: int64
        - name: date_from
          in: query
          description: Start date filter (inclusive). Format `YYYY-MM-DD` or ISO 8601.
          required: false
          schema:
            type: string
            format: date-time
        - name: date_to
          in: query
          description: End date filter (inclusive). Format `YYYY-MM-DD` or ISO 8601.
          required: false
          schema:
            type: string
            format: date-time
        - name: orderby
          in: query
          description: Field to order results by.
          required: false
          schema:
            type: string
            enum: [id, event_type, post_id, duration_ms, logged_at]
            default: logged_at
        - name: order
          in: query
          description: Sort direction.
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: A paginated list of performance log entries.
          headers:
            X-WP-Total:
              $ref: '#/components/headers/X-WP-Total'
            X-WP-TotalPages:
              $ref: '#/components/headers/X-WP-TotalPages'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PerformanceLogEntry'
              example:
                - id: 89201
                  event_type: process_post
                  post_id: 8234
                  duration_ms: 210
                  memory_kb: 9400
                  rules_checked: 1024
                  links_applied: 4
                  logged_at: "2026-03-24T10:15:02"
                - id: 89200
                  event_type: process_post
                  post_id: 7012
                  duration_ms: 145
                  memory_kb: 7800
                  rules_checked: 1024
                  links_applied: 2
                  logged_at: "2026-03-24T10:14:58"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

  # ---------------------------------------------------------------------------
  # Cache
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        WordPress Application Passwords. The username is your WordPress
        username and the password is an Application Password generated at
        **Users > Profile > Application Passwords**.

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
    PageParam:
      name: page
      in: query
      description: Current page of the collection.
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      description: Maximum number of items to return per page.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  headers:
    X-WP-Total:
      description: Total number of items in the collection.
      schema:
        type: integer
    X-WP-TotalPages:
      description: Total number of pages in the collection.
      schema:
        type: integer

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    AppliedLink:
      type: object
      description: A link that was inserted into a post by the engine.
      required:
        - id
        - post_id
        - rule_id
        - keyword
        - target_url
        - applied_at
      properties:
        id:
          type: integer
          format: int64
          description: Unique applied link identifier.
          readOnly: true
          example: 10234
        post_id:
          type: integer
          format: int64
          description: The WordPress post ID where this link was applied.
          example: 8234
        rule_id:
          type: integer
          format: int64
          description: The rule ID that generated this link.
          example: 1
        keyword:
          type: string
          maxLength: 255
          description: The keyword that was matched and linked.
          example: inteligencia artificial
        target_url:
          type: string
          format: uri
          description: The URL the keyword links to.
          example: https://example.com/glosario/inteligencia-artificial/
        applied_at:
          type: string
          format: date-time
          description: When this link was applied to the post.
          example: "2026-03-24T10:15:02"
    AppliedStats:
      type: object
      description: Aggregated statistics about applied links across all posts.
      properties:
        total_links:
          type: integer
          description: Total number of applied links across all posts.
          example: 48210
        by_rule_type:
          type: object
          description: Link count grouped by rule type.
          properties:
            internal:
              type: integer
              example: 28430
            entity:
              type: integer
              example: 15200
            affiliate:
              type: integer
              example: 4580
        posts_with_links:
          type: integer
          description: Number of distinct posts that have at least one applied link.
          example: 14320
        avg_links_per_post:
          type: number
          format: float
          description: Average number of applied links per post (across posts with links).
          example: 3.37
        top_linked_posts:
          type: array
          description: Posts with the highest number of applied links (top 10).
          maxItems: 10
          items:
            type: object
            properties:
              post_id:
                type: integer
                format: int64
                example: 1204
              post_title:
                type: string
                example: "Guia completa de inteligencia artificial para startups"
              link_count:
                type: integer
                example: 12
        top_rules:
          type: array
          description: Rules with the highest application count (top 10).
          maxItems: 10
          items:
            type: object
            properties:
              rule_id:
                type: integer
                format: int64
                example: 1
              keyword:
                type: string
                example: inteligencia artificial
              applied_count:
                type: integer
                example: 4230

    # --- Exclusions ----------------------------------------------------------
...
    PerformanceLogEntry:
      type: object
      description: A single entry in the performance log tracking one processing event.
      required:
        - id
        - event_type
        - logged_at
      properties:
        id:
          type: integer
          format: int64
          description: Unique log entry identifier.
          readOnly: true
          example: 89201
        event_type:
          type: string
          maxLength: 100
          description: |
            The type of event logged:
            - `process_post`: Successful post processing.
            - `process_failed`: Failed post processing.
            - `cache_hit`: Content served from cache on the frontend.
            - `cache_miss`: Cache miss on the frontend (original content served).
            - `cache_flush`: Cache flush event.
            - `cache_warm`: Cache warming event.
            - `bulk_start`: Bulk reprocess started.
            - `bulk_complete`: Bulk reprocess completed.
          example: process_post
        post_id:
          type: integer
          format: int64
          nullable: true
          description: The WordPress post ID associated with this event. Null for global events.
          example: 8234
        duration_ms:
          type: integer
          nullable: true
          description: Processing duration in milliseconds.
          example: 210
        memory_kb:
          type: integer
          nullable: true
          description: Peak memory usage in kilobytes during processing.
          example: 9400
        rules_checked:
          type: integer
          nullable: true
          description: Number of rules evaluated during processing.
          example: 1024
        links_applied:
          type: integer
          nullable: true
          description: Number of links inserted into the post content.
          example: 4
        logged_at:
          type: string
          format: date-time
          description: When this event was logged.
          example: "2026-03-24T10:15:02"
    PerformanceSummary:
      type: object
      description: Aggregated performance metrics over a time period.
      properties:
        period:
          type: string
          description: The time period these metrics cover.
          example: "24h"
        processing:
          type: object
          description: Post processing metrics.
          properties:
            total_posts_processed:
              type: integer
              description: Total posts processed in the period.
              example: 1420
            avg_duration_ms:
              type: number
              format: float
              description: Average processing time per post in milliseconds.
              example: 187.4
            p95_duration_ms:
              type: number
              format: float
              description: 95th percentile processing time in milliseconds.
              example: 432.0
            max_duration_ms:
              type: number
              format: float
              description: Maximum processing time observed in the period.
              example: 890.0
            avg_rules_checked:
              type: integer
              description: Average number of rules evaluated per post.
              example: 1024
            avg_links_applied:
              type: number
              format: float
              description: Average number of links applied per post.
              example: 3.2
            avg_memory_kb:
              type: integer
              description: Average peak memory usage per post in kilobytes.
              example: 8420
            max_memory_kb:
              type: integer
              description: Maximum memory usage observed in the period.
              example: 24100
        cache:
          type: object
          description: Cache performance metrics.
          properties:
            hit_rate:
              type: number
              format: float
              minimum: 0
              maximum: 1
              description: Cache hit rate as a decimal (0.0 to 1.0).
              example: 0.94
            hits:
              type: integer
              description: Total cache hits in the period.
              example: 45230
            misses:
              type: integer
              description: Total cache misses in the period.
              example: 2870
            estimated_memory_mb:
              type: number
              format: float
              description: Estimated memory used by LeanAutoLinks in the object cache backend.
              example: 148.5
        queue:
          type: object
          description: Current queue status (snapshot, not aggregated).
          properties:
            pending:
              type: integer
              example: 12
            processing:
              type: integer
              example: 3
            done:
              type: integer
              example: 14832
            failed:
              type: integer
              example: 2
            estimated_drain_seconds:
              type: integer
              description: Estimated seconds to process all pending items based on current throughput.
              example: 45
        throughput:
          type: object
          description: Processing throughput metrics.
          properties:
            posts_per_minute:
              type: number
              format: float
              description: Current throughput in posts processed per minute.
              example: 14.8
            sustained_posts_per_hour:
              type: integer
              description: Sustained throughput extrapolated to posts per hour.
              example: 888

    # --- Cache ---------------------------------------------------------------
...
    Error:
      type: object
      description: |
        Standard error response following WordPress REST API conventions.
      required:
        - code
        - message
        - data
      properties:
        code:
          type: string
          description: |
            Machine-readable error code. Common codes:
            - `rest_forbidden`: Missing required capability.
            - `rest_no_route`: Endpoint not found.
            - `invalid_parameter`: A request parameter failed validation.
            - `missing_parameter`: A required parameter is missing.
            - `not_found`: The requested resource does not exist.
            - `duplicate_keyword`: A rule with this keyword already exists.
            - `internal_error`: An unexpected server error occurred.
          example: invalid_parameter
        message:
          type: string
          description: Human-readable error description.
          example: "The rule_type parameter must be one of: internal, affiliate, entity."
        data:
          type: object
          description: Additional error context.
          properties:
            status:
              type: integer
              description: The HTTP status code.
              example: 400
            param:
              type: string
              description: The parameter that caused the error (if applicable).
              example: rule_type
            details:
              type: object
              description: Additional error details (structure varies by error type).

    # --- Pagination ----------------------------------------------------------
...
  responses:
    BadRequest:
      description: |
        The request is malformed or contains invalid parameters. Check the
        `code` and `data.param` fields to identify the issue.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: invalid_parameter
            message: "The rule_type parameter must be one of: internal, affiliate, entity."
            data:
              status: 400
              param: rule_type
    Unauthorized:
      description: |
        Authentication is missing or invalid. Ensure you are sending a valid
        WordPress Application Password via HTTP Basic Auth.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: rest_not_logged_in
            message: "You are not currently logged in."
            data:
              status: 401
    Forbidden:
      description: |
        The authenticated user does not have the `manage_options` capability
        required to access this endpoint.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: rest_forbidden
            message: "Sorry, you are not allowed to manage LeanAutoLinks settings."
            data:
              status: 403
...
    InternalError:
      description: An unexpected server error occurred. Check the WordPress error log for details.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: internal_error
            message: "An unexpected error occurred. Please try again later."
            data:
              status: 500